+++ src/actions/vocabularies/vocabularies.js
@@ -7,7 +7,6 @@ import {
   GET_VOCABULARY,
   GET_VOCABULARY_TOKEN_TITLE,
 } from '@plone/volto/constants/ActionTypes';
-import config from '@plone/volto/registry';
 
 /**
  * Get vocabulary given a URL (coming from a Schema) or from a vocabulary name.
@@ -18,11 +17,10 @@ import config from '@plone/volto/registry';
  * @returns {Object} Get vocabulary action.
  */
 export function getVocabulary(vocabNameOrURL, query = null, start = 0) {
-  const { settings } = config;
   // In case we have a URL, we have to get the vocabulary name
-  const vocabulary =
-    vocabNameOrURL &&
-    vocabNameOrURL.replace(`${settings.apiPath}/@vocabularies/`, '');
+  const vocabulary = /(?<=\/@vocabularies\/)(.*)/.test(vocabNameOrURL)
+    ? /(?<=\/@vocabularies\/)(.*)/.exec(vocabNameOrURL)[0]
+    : vocabNameOrURL;
   let queryString = `b_start=${start}`;
   if (query) {
     queryString = `${queryString}&title=${query}`;
@@ -47,12 +45,10 @@ export function getVocabulary(vocabNameOrURL, query = null, start = 0) {
  * @returns {Object} Get vocabulary action.
  */
 export function getVocabularyTokenTitle(vocabNameOrURL, token = null) {
-  const { settings } = config;
   // In case we have a URL, we have to get the vocabulary name
-  const vocabulary = vocabNameOrURL.replace(
-    `${settings.apiPath}/@vocabularies/`,
-    '',
-  );
+  const vocabulary = /(?<=\/@vocabularies\/)(.*)/.test(vocabNameOrURL)
+    ? /(?<=\/@vocabularies\/)(.*)/.exec(vocabNameOrURL)[0]
+    : vocabNameOrURL;
 
   return {
     type: GET_VOCABULARY_TOKEN_TITLE,