+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ### Internal
 
+- Add `applySchemaDefaults`, in addition to `applyBlockDefaults`, to allow reuse in object widgets and other advanced scenarios @tiberiuichim
+
 ## 14.0.0-alpha.35 (2021-11-24)
 
 ### Bugfix