+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ### Internal
 
+- Fix references to old configuration style in apiExpanders documentation @tiberiuichim
 - Add `applySchemaDefaults`, in addition to `applyBlockDefaults`, to allow reuse in object widgets and other advanced scenarios @tiberiuichim
 
 ## 14.0.0-alpha.35 (2021-11-24)